upgrade to holochain 0.6.0

+++ src/decode_webapp.rs
@@ -5,6 +5,7 @@ use sha2::{Digest, Sha256};
 use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
+
 #[napi(object)]
 pub struct HappAndUiHashes {
     pub happ_sha256: String,
@@ -27,13 +28,13 @@ pub async fn happ_bytes_with_custom_properties(
 ) -> napi::Result<Vec<u8>> {
     // 1. read and decode happ bundle
     let happ_bytes = fs::read(happ_path)?;
-    let app_bundle = AppBundle::decode(&happ_bytes)
+    let app_bundle = AppBundle::unpack(happ_bytes.as_slice())
         .map_err(|e| napi::Error::from_reason(format!("Failed to decode happ file: {}", e)))?;
 
     let mut manifest = app_bundle.manifest().clone();
 
     match manifest {
-        AppManifest::V1(ref mut m) => {
+        AppManifest::V0(ref mut m) => {
             for role_manifest in m.roles.iter_mut() {
                 match properties.get(&role_manifest.name) {
                     Some(maybe_props) => match maybe_props {
@@ -67,14 +68,14 @@ pub async fn happ_bytes_with_custom_properties(
 
     let modified_app_bundle = AppBundle::from(modified_bundle);
 
-    let app_bundle_bytes = modified_app_bundle.encode().map_err(|e| {
+    let app_bundle_bytes = modified_app_bundle.into_inner().pack().map_err(|e| {
         napi::Error::from_reason(format!(
             "Failed to encode modified AppBundle to bytes: {}",
             e
         ))
     })?;
 
-    Ok(app_bundle_bytes)
+    Ok(app_bundle_bytes.into())
 }
 
 /// Saves a happ or a webhapp file. If a uis_dir is specified and it is a webhapp,
@@ -88,7 +89,7 @@ pub async fn save_happ_or_webhapp(
 ) -> napi::Result<StoredHappPathAndHashes> {
     let happ_or_webhapp_bytes = fs::read(happ_or_web_happ_path)?;
 
-    match WebAppBundle::decode(&happ_or_webhapp_bytes) {
+    match WebAppBundle::unpack(happ_or_webhapp_bytes.as_slice()) {
         Ok(web_app_bundle) => {
             let web_happ_hash = hash_bytes_sha256(happ_or_webhapp_bytes);
 
@@ -97,7 +98,7 @@ pub async fn save_happ_or_webhapp(
                 napi::Error::from_reason(format!("Failed to extract ui zip bytes: {}", e))
             })?;
 
-            let ui_hash = hash_bytes_sha256(web_ui_zip_bytes.clone().into_owned().into_inner());
+            let ui_hash = hash_bytes_sha256(web_ui_zip_bytes.clone().into());
 
             // extracting happ bundle
             let app_bundle = web_app_bundle.happ_bundle().await.map_err(|e| {
@@ -107,11 +108,11 @@ pub async fn save_happ_or_webhapp(
                 ))
             })?;
 
-            let app_bundle_bytes = app_bundle.encode().map_err(|e| {
+            let app_bundle_bytes = app_bundle.into_inner().pack().map_err(|e| {
                 napi::Error::from_reason(format!("Failed to encode happ to bytes: {}", e))
             })?;
 
-            let happ_hash = hash_bytes_sha256(app_bundle_bytes);
+            let happ_hash = hash_bytes_sha256(app_bundle_bytes.clone().into());
 
             let happ_path = PathBuf::from(happs_dir).join(format!("{}.happ", happ_hash));
             let happ_path_string = happ_path
@@ -121,43 +122,36 @@ pub async fn save_happ_or_webhapp(
                 .map_err(|e| napi::Error::from_reason(e))?;
 
             // Store UI if uis_dir is specified
-            match uis_dir {
-                Some(dir) => {
-                    let ui_target_dir = PathBuf::from(dir).join(ui_hash.clone()).join("assets");
-                    if !path_exists(&ui_target_dir) {
-                        fs::create_dir_all(&ui_target_dir)?;
-                    }
-
-                    let ui_zip_path = PathBuf::from(ui_target_dir.clone()).join("ui.zip");
-
-                    // unzip and store UI
-                    fs::write(
-                        ui_zip_path.clone(),
-                        web_ui_zip_bytes.into_owned().into_inner(),
-                    )
-                    .map_err(|e| {
-                        napi::Error::from_reason(format!("Failed to write Web UI Zip file: {}", e))
-                    })?;
-
-                    let file = fs::File::open(ui_zip_path.clone()).map_err(|e| {
-                        napi::Error::from_reason(format!("Failed to read Web UI Zip file: {}", e))
-                    })?;
-
-                    unzip_file(file, ui_target_dir.into()).map_err(|e| {
-                        napi::Error::from_reason(format!("Failed to unzip ui.zip: {}", e))
-                    })?;
-
-                    fs::remove_file(ui_zip_path).map_err(|e| {
-                        napi::Error::from_reason(format!(
-                            "Failed to remove ui.zip after unzipping: {}",
-                            e
-                        ))
-                    })?;
+            if let Some(dir) = uis_dir {
+                let ui_target_dir = PathBuf::from(dir).join(ui_hash.clone()).join("assets");
+                if !path_exists(&ui_target_dir) {
+                    fs::create_dir_all(&ui_target_dir)?;
                 }
-                None => (),
+
+                let ui_zip_path = PathBuf::from(ui_target_dir.clone()).join("ui.zip");
+
+                // unzip and store UI
+                fs::write(ui_zip_path.clone(), &web_ui_zip_bytes).map_err(|e| {
+                    napi::Error::from_reason(format!("Failed to write Web UI Zip file: {}", e))
+                })?;
+
+                let file = fs::File::open(ui_zip_path.clone()).map_err(|e| {
+                    napi::Error::from_reason(format!("Failed to read Web UI Zip file: {}", e))
+                })?;
+
+                unzip_file(file, ui_target_dir.into()).map_err(|e| {
+                    napi::Error::from_reason(format!("Failed to unzip ui.zip: {}", e))
+                })?;
+
+                fs::remove_file(ui_zip_path).map_err(|e| {
+                    napi::Error::from_reason(format!(
+                        "Failed to remove ui.zip after unzipping: {}",
+                        e
+                    ))
+                })?;
             }
 
-            app_bundle.write_to_file(&happ_path).await.map_err(|e| {
+            fs::write(&happ_path, app_bundle_bytes).map_err(|e| {
                 napi::Error::from_reason(format!("Failed to write .happ file: {}", e))
             })?;
 
@@ -169,15 +163,15 @@ pub async fn save_happ_or_webhapp(
             })
         }
         Err(_) => {
-            let app_bundle = AppBundle::decode(&happ_or_webhapp_bytes).map_err(|e| {
+            let app_bundle = AppBundle::unpack(happ_or_webhapp_bytes.as_slice()).map_err(|e| {
                 napi::Error::from_reason(format!("Failed to decode happ file: {}", e))
             })?;
 
-            let app_bundle_bytes = app_bundle.encode().map_err(|e| {
+            let app_bundle_bytes = app_bundle.into_inner().pack().map_err(|e| {
                 napi::Error::from_reason(format!("Failed to encode happ to bytes: {}", e))
             })?;
 
-            let happ_hash = hash_bytes_sha256(app_bundle_bytes);
+            let happ_hash = hash_bytes_sha256(app_bundle_bytes.clone().into());
 
             let happ_path = PathBuf::from(happs_dir).join(format!("{}.happ", happ_hash));
             let happ_path_string = happ_path
@@ -186,7 +180,7 @@ pub async fn save_happ_or_webhapp(
                 .ok_or("Failed to convert happ path to string")
                 .map_err(|e| napi::Error::from_reason(e))?;
 
-            app_bundle.write_to_file(&happ_path).await.map_err(|e| {
+            fs::write(&happ_path, &app_bundle_bytes).map_err(|e| {
                 napi::Error::from_reason(format!("Failed to write .happ file: {}", e))
             })?;
             Ok(StoredHappPathAndHashes {
@@ -205,42 +199,44 @@ pub async fn save_happ_or_webhapp(
 pub async fn validate_happ_or_webhapp(
     happ_or_webhapp_bytes: Vec<u8>,
 ) -> napi::Result<HappAndUiHashes> {
-    let (app_bundle, maybe_ui_and_webhapp_hash) = match WebAppBundle::decode(&happ_or_webhapp_bytes)
-    {
-        Ok(web_app_bundle) => {
-            let mut hasher = Sha256::new();
-            hasher.update(happ_or_webhapp_bytes);
-            let web_happ_hash = hex::encode(hasher.finalize());
-            // extracting ui.zip bytes
-            let web_ui_zip_bytes = web_app_bundle.web_ui_zip_bytes().await.map_err(|e| {
-                napi::Error::from_reason(format!("Failed to extract ui zip bytes: {}", e))
-            })?;
+    let (app_bundle, maybe_ui_and_webhapp_hash) =
+        match WebAppBundle::unpack(happ_or_webhapp_bytes.as_slice()) {
+            Ok(web_app_bundle) => {
+                let mut hasher = Sha256::new();
+                hasher.update(happ_or_webhapp_bytes);
+                let web_happ_hash = hex::encode(hasher.finalize());
+                // extracting ui.zip bytes
+                let web_ui_zip_bytes = web_app_bundle.web_ui_zip_bytes().await.map_err(|e| {
+                    napi::Error::from_reason(format!("Failed to extract ui zip bytes: {}", e))
+                })?;
 
-            let mut hasher = Sha256::new();
-            hasher.update(web_ui_zip_bytes.clone().into_owned().into_inner());
-            let ui_hash = hex::encode(hasher.finalize());
+                let mut hasher = Sha256::new();
+                hasher.update(web_ui_zip_bytes.clone());
+                let ui_hash = hex::encode(hasher.finalize());
 
-            // extracting happ bundle
-            let app_bundle = web_app_bundle.happ_bundle().await.map_err(|e| {
-                napi::Error::from_reason(format!(
-                    "Failed to get happ bundle from webapp bundle bytes: {}",
-                    e
-                ))
-            })?;
+                // extracting happ bundle
+                let app_bundle = web_app_bundle.happ_bundle().await.map_err(|e| {
+                    napi::Error::from_reason(format!(
+                        "Failed to get happ bundle from webapp bundle bytes: {}",
+                        e
+                    ))
+                })?;
 
-            (app_bundle, Some((ui_hash, web_happ_hash)))
-        }
-        Err(_) => {
-            let app_bundle = AppBundle::decode(&happ_or_webhapp_bytes).map_err(|e| {
-                napi::Error::from_reason(format!("Failed to decode happ file: {}", e))
-            })?;
-            (app_bundle, None)
-        }
-    };
+                (app_bundle, Some((ui_hash, web_happ_hash)))
+            }
+            Err(_) => {
+                let app_bundle =
+                    AppBundle::unpack(happ_or_webhapp_bytes.as_slice()).map_err(|e| {
+                        napi::Error::from_reason(format!("Failed to decode happ file: {}", e))
+                    })?;
+                (app_bundle, None)
+            }
+        };
 
     let mut hasher = Sha256::new();
     let app_bundle_bytes = app_bundle
-        .encode()
+        .into_inner()
+        .pack()
         .map_err(|e| napi::Error::from_reason(format!("Failed to encode happ to bytes: {}", e)))?;
     hasher.update(app_bundle_bytes);
     let happ_hash = hex::encode(hasher.finalize());
